move WorkerPool into db module
The pool is independent of the indexer and may also be used
by other parts of the software.
